2.1-7pre20 Oct 16 2004                 x86_64 Linux Fixes fixes BlackBox Serial now runs                 switch to JNI_VERSION_1_2 -- many people are still using                 1.3 JREs

+++ src/ParallelImp.c
@@ -575,7 +575,7 @@ JNIEXPORT jint JNICALL LPRPort(readArray)( JNIEnv *env,
 	threshold = get_java_var( env, jobj,"threshold","I" );
 	timeout = get_java_var( env, jobj,"threshold","I" );
 
-	if( length < 1 || length > SSIZE_MAX )
+	if( (size_t) length < 1 || (size_t) length > SSIZE_MAX )
 	{
 		throw_java_exception( env, IO_EXCEPTION, "readArray",
 			"Invalid length" );
